Extract Intcode from Day 9
